memecoin_scout/requirements.txt
httpx>=0.24
pydantic>=2.6
pyyaml>=6.0
rich>=13.7
